refactor: move sorting component to separate file
