+++ src/view/sorting.js
@@ -1,0 +1,11 @@
+export {renderSortComponent};
+
+const renderSortComponent = () => {
+  return `
+    <ul class="sort">
+      <li><a href="#" class="sort__button sort__button--active">Sort by default</a></li>
+      <li><a href="#" class="sort__button">Sort by date</a></li>
+      <li><a href="#" class="sort__button">Sort by rating</a></li>
+    </ul>
+  `;
+};
